+++ .gitignore
@@ -51,10 +51,12 @@ coverage.xml
 *.mo
 *.pot
 
-# Django stuff:
+# Webclient stuff:
 *.log
 db.sqlite3
 .sass-cache
+bundles/
+node_modules/
 
 # Sphinx documentation
 docs/_build/
