+++ clients/webserver/deckr/README.md
@@ -1,0 +1,13 @@
+# Deckr Webclient
+
+## Requirements
+To run the webclient, you need to install the following:
+
+- [PhantomJS](http://phantomjs.org/)
+- [NPM](https://www.npmjs.com/)
+- [Compass](http://compass-style.org/)
+
+## Installation
+To compile the JSX for the game room, run `./node_modules/.bin/webpack --config webpack.config.js`
+
+If you wish to make changes to the stylesheets, run `compass watch`
