+++ clients/webserver/deckr/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "deckr",
+  "version": "1.0.0",
+  "description": "Deckr Client",
+  "author": "Alexander Bake",
+  "license": "ISC",
+  "devDependencies": {
+    "babel": "^5.8.23",
+    "babel-core": "^5.8.25",
+    "babel-loader": "^5.3.2",
+    "react": "^0.13.3",
+    "webpack": "^1.12.2",
+    "webpack-bundle-tracker": "0.0.8"
+  }
+}
