+++ clients/webserver/deckr/webclient/templates/webclient/base.html
@@ -6,8 +6,6 @@
     {% load staticfiles %}
     <!-- JS -->
     <script src="//code.jquery.com/jquery-2.1.4.min.js"></script>
-    <script src="{% static 'js/react/react.js' %}"></script>
-    <script src="{% static 'js/react/JSXTransformer.js' %}"></script>
     <script src="{% static 'js/sockjs/sockjs.js' %}"></script>
     <script src="{% static 'js/protobuf/Long.js' %}"></script>
     <script src="{% static 'js/protobuf/ByteBuffer.js' %}"></script>
