+++ clients/webserver/deckr/webpack-stats.json
@@ -1,0 +1,1 @@
+{"status":"done","chunks":{"main":[{"name":"main-e6d31320ae174000807d.js","path":"/Users/abake/Desktop/personal_projects/deckr2/clients/webserver/deckr/webclient/static/bundles/main-e6d31320ae174000807d.js"}]}}
